Subimos Prueba_PWM1.
Este firmware hace funcionar el PWM en 50KHz, con saltos de ton de los
canales PWM0 y PWM1 entre los siguientes valores: 2, 5, 10 y 18uS. Los
mismos se obtienen presionando IRQ_SW2.
La idea es que sirva como base para ver el funcionamiento de este
periférico.

+++ Firmware/Test/Control_Tension_ADC_PWM/Project_Headers/Leds.h
@@ -8,11 +8,16 @@
 #ifndef LEDS_H_
 #define LEDS_H_
 
+#include "Cpu.h"
+
 /* Variables externas */
-extern uint16_t ledEstado;
-extern uint16_t borrarContador;
+//extern uint16_t ledEstado;
+//extern uint16_t borrarContador;
+extern unsigned int ledEstado;
+extern unsigned int borrarContador;
 
 /* Prototipos de función */
-void Led_Estado(uint16_t estado);
+//void Led_Estado(uint16_t estado);
+void Led_Estado(unsigned int estado);
 
 #endif /* LEDS_H_ */
